+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to the "Znuny (DK4Znuny-VisualStudioCode)" extension will be documented in this file.
 
+## [1.0.3]
+
+- ⭐ StatusBar
+  - The status bar gets an additional **Znuny** item and the entire status bar is displayed in the Znuny color ![#ff9b00](https://placehold.co/15x15/ff9b00/ff9b00.png) if the active file is a "Znuny file".
+    - If the active file belongs to a **Znuny framework** ("Znuny file") in the workspace, the **product name** and **version** are displayed from the RELEASE file.
+    - If the active file belongs to a **Znuny package** ("Znuny file") in the workspace, the **vendor** and the largest **framework version** are displayed from the SOPM file.
+- Snippets
+  - Fixed - Broken snippet - Functions/TicketObject/znuny.TicketObject.TicketSearch.code-snippets [#2](https://github.com/dennykorsukewitz/DK4Znuny-VisualStudioCode/issues/2)
+  - Added missing DiffObject snippet for 6.0
+  - Tidied Snippets
+
 ## [1.0.2]
 
 - Snippets
@@ -18,12 +29,12 @@ All notable changes to the "Znuny (DK4Znuny-VisualStudioCode)" extension will be
   - Fixed syntax typos
   - Tidied snippet filenames for better readability
   - Added DynamicFields
-    - WebserviceText        (⭐ new)
-    - WebserviceMultiselect (⭐ new)
-    - WebserviceDropdown    (⭐ new)
+    - ⭐ WebserviceText
+    - ⭐ WebserviceMultiselect
+    - ⭐ WebserviceDropdown
   - Added SeleniumObject
-    - VerifiedClick         (⭐ new)
-    - VerifiedSubmit        (⭐ new)
+    - ⭐ VerifiedClick
+    - ⭐ VerifiedSubmit
 
 ## [1.0.0]